requirements.txt
Flask==2.2.2
google-cloud-datastore==2.8.2
requests==2.28.1
google-auth==2.14.1 
google-auth-oauthlib==0.7.1 
oauthlib==3.2.2 
requests-oauthlib==1.3.1
google-api-python-client==2.65.0
google-auth-httplib2==0.1.0
httplib2==0.21.0 
uritemplate==4.1.1
json2html==1.3.0